+++ apps/zhi-formdata-fetch/src/index.ts
@@ -1,3 +1,6 @@
 import FormdataFetch from "./lib/formdataFetch"
+import fetch, { Blob, FormData, Headers } from "node-fetch"
 
 export { FormdataFetch }
+export { fetch }
+export { Blob, FormData, Headers }