+++ apps/zhi-formdata-fetch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-formdata-fetch
 
+## 0.2.0
+
+### Minor Changes
+
+- feat: change form fetch
+
+## 0.1.3
+
+### Patch Changes
+
+- feat: export formDataToBlob
+
 ## 0.1.2
 
 ### Patch Changes